+++ scripts/lichessTools-init.js
@@ -72,6 +72,7 @@ lichessTools.loadTool(LiChessTools.Tools.PlayerLag);
 lichessTools.loadTool(LiChessTools.Tools.ProfileSliderText);
 lichessTools.loadTool(LiChessTools.Tools.FixCoords);
 lichessTools.loadTool(LiChessTools.Tools.DownloadAllStudies);
+lichessTools.loadTool(LiChessTools.Tools.ProfilePuzzleTab);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);